+++ ics23/proof.go
@@ -1,4 +1,4 @@
-package iavl
+package ics23
 
 import (
 	"fmt"
